BUG: incorrect row/column, revert division change
Futurize was used to automatically convert the division
operator / for python3 compativility [1].  This was done
because Python2 had different behavior for int vs float
arguments in terms of rounding behavior [2].

When the future package was deprectded as slicer
mpved to pythhon3, the old_div calls were mapped
to the // operator [3].

However in the row/column case, the sqrt operator
could be float, and the // flooring division
behavior led to incorrect calculations.

Other places, where pixel space is expected
integer division with // is used.  Where float
values are expected / is used.

[1] https://github.com/pieper/CompareVolumes/commit/e671f113f89a03b61103139b9037b02f1e17877e

[2] https://python-future.org/compatible_idioms.html#division

[3] https://github.com/pieper/CompareVolumes/commit/fa7d059af16e765f806d459666d3809a00f06add